requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Machine Learning
torch==2.1.0
torchvision==0.16.0
torchxrayvision==1.0.1
numpy==1.24.3
pillow==10.0.1
opencv-python==4.8.1.78
scikit-image==0.21.0

# Medical imaging
pydicom==2.4.3
SimpleITK==2.3.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For advanced preprocessing
nibabel==5.1.0
matplotlib==3.8.2
